hook get tab status added
