+++ src/components/navigation/searchBar/SearchBar.tsx
@@ -24,6 +24,7 @@ const SearchBar: React.FC = () => {
     right: null,
   });
 
+  // TODO: coordinates only top / and centered
   useEffect(() => {
     if (inputRef.current) inputRef.current.focus();
