+++ src/components/searchList/SearchList.tsx
@@ -1,8 +1,5 @@
 import { ICocktail } from 'types/cocktail';
 import { IIngredient } from 'types/ingredient';
-// import BarList from 'components/barList';
-// import CocktailCard from 'components/cocktailList/cocktailCard';
-// import { ListItem } from 'components/cocktailList/CocktailList.styled';
 import CocktailList from 'components/cocktailList';
 import IngredientsList from 'components/ingredientsList';
